+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 #3.2.4
 - Fixed issue with tooltip not showing up on triple click
 - Disabled accordeon-like animation on custom search panel reveal
-- Added config to exclude domains for 'Snap selection to words'
+- Added config to exclude domains for 'Snap selection to words' function
+- Tooltip now will not be recreated when user clicks on link or button - and there's selected text somewhere else on page 
 
 #3.2.3
 - Small layout fixes and improvements